A few more non-substantive changes to get this beast working on ppc. The only thing I'd be at all worried about is the (long) cast I removed in tcp_input.c. It's line 507 or so, and I just commented it out, as it didn't seem to serve a useful function, and mwcc barfed on it.

+++ net_kit/source/server/test/timer_test.c
@@ -41,7 +41,7 @@ int main(void)
 	gStart = real_time_clock();
 
 	for (i = 0;i < TIMERS;i++)
-		timer[i] = net_add_timer(hook,(void *)i + 1,gIntervals[i] * 1000000LL);
+		timer[i] = net_add_timer(hook,(void *)((uint32)i + 1UL),gIntervals[i] * 1000000LL);
 
 	puts("snoozing...");
 	snooze(SNOOZE * 1000000LL);
